+++ src/components/LinkApp.tsx
@@ -19,6 +19,7 @@ import { textVariants } from '@theme';
 
 type StyleProps = Omit<TextProps, 'textTransform'> & {
   $textTransform?: TextProps['textTransform'];
+  $animate?: boolean;
 };
 
 const SLink = styled.a<StyleProps & HTMLAnchorElement>`
@@ -33,6 +34,13 @@ const SLink = styled.a<StyleProps & HTMLAnchorElement>`
   ${typography}
   ${layout}
   ${({ $textTransform }) => $textTransform && { 'text-transform': $textTransform }}
+  ${({ $animate }) =>
+    $animate &&
+    `&:hover {
+      transform: scale(1.05);
+      transition: all 300ms;
+    }
+    transition: all 300ms;`}
 `;
 
 const SRouterLink = styled(RouterLink)<StyleProps & RouterLinkProps>`
@@ -47,6 +55,13 @@ const SRouterLink = styled(RouterLink)<StyleProps & RouterLinkProps>`
   ${typography}
   ${layout}
   ${({ $textTransform }) => $textTransform && { 'text-transform': $textTransform }}
+  ${({ $animate }) =>
+    $animate &&
+    `&:hover {
+      transform: scale(1.05);
+      transition: all 300ms;
+    }
+    transition: all 300ms;`}
 `;
 
 interface LinkProps {
@@ -62,6 +77,7 @@ const LinkApp: React.FC<LinkAppProps & StyleProps> = ({
   href,
   isExternal = false,
   variant = 'defaultLink',
+  $animate = false,
   onClick,
   ...props
 }) => {
@@ -70,6 +86,7 @@ const LinkApp: React.FC<LinkAppProps & StyleProps> = ({
       `LinkApp: Received href prop ${href}. Set prop isExternal to use an external link.`
     );
   }
+
   const handleClick = (event: React.MouseEvent<HTMLAnchorElement, MouseEvent>) => {
     if (!onClick) return;
     event.preventDefault();
@@ -80,13 +97,14 @@ const LinkApp: React.FC<LinkAppProps & StyleProps> = ({
     <SLink
       href={href}
       variant={variant}
+      $animate={$animate}
       target="_blank"
       rel="noreferrer"
       onClick={handleClick}
       {...props}
     />
   ) : (
-    <SRouterLink to={href} variant={variant} onClick={handleClick} {...props} />
+    <SRouterLink to={href} variant={variant} $animate={$animate} onClick={handleClick} {...props} />
   );
 };
