+++ src/features/Layout/Navigation/components/index.ts
@@ -3,7 +3,7 @@ export { NavLink } from './NavLink';
 export { Tray } from './Tray';
 export { TrayItem } from './TrayItem';
 export { NavTray } from './NavTray';
-export { ExternalLink } from './ExternalLink';
+export { TrayLink } from './TrayLink';
 export { SupportUsTray } from './SupportUsTray';
 export { TopItem } from './TopItem';
 export { Subscribe } from './Subscribe';
